+++ bifrostlib/__init__.py
@@ -4,4 +4,5 @@ __all__ = [
     'database_interface'
     ]
 
+__version__ = '2.1.1'